Sidebar components code cleanup
