+++ frontend/src/components/SidebarTree/sidebar_tree.js
@@ -3,12 +3,11 @@ import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
 import Tree from 'react-file-view'
-import { selectFile, addFile, renameFile, setExamples } from '../../actions/index'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+import { selectFile, setExamples } from '../../actions/index'
 import pathParser from './path_parser'
 import styles from './sidebar_tree.css'
-import theme from '../../theme/theme.css'
 import { FETCH_EXAMPLE_LIST_URL } from '../../constants'
+import SidebarTreeItem from '../SidebarTreeItem'
 
 class SidebarTree extends Component {
   componentDidMount() {
@@ -25,135 +24,13 @@ class SidebarTree extends Component {
       })
   }
 
-  renderNode = node => {
-    let fileIcon
-    let nodeFolder = false
-    if (node.leaf) {
-      fileIcon = 'file-code'
-    } else if (node.collapsed) {
-      fileIcon = 'folder'
-    } else {
-      fileIcon = 'folder-open'
-      nodeFolder = true
-    }
-
-    let iconClass = [styles.icon, theme.sidebarIcon].join(' ')
-    let nodeActive = false
-    if (node === this.props.files.active) {
-      nodeActive = true
-      iconClass = [styles.icon, theme.sidebarActiveIcon].join(' ')
-    }
-
-    return (
-      <span
-        onClick={() => this.fileClick(node)}
-        className={styles.sidebarItem}
-      >
-        <FontAwesomeIcon className={iconClass} icon={fileIcon} />
-        <span className={[styles.sidebarItemTitle, theme.sidebarItemText].join(' ')}>
-          {node.module}
-        </span>
-        {/*
-          Show dot icon next to the node that is currently selected.
-        */}
-        {
-          <FontAwesomeIcon
-            className={[styles.circleIcon, theme.sidebarActiveIcon].join(' ')}
-            icon='dot-circle'
-            size='xs'
-            style={nodeActive ? { display: 'block' } : { display: 'none' }}
-          />
-        }
-        {/*
-          If node is a folder show additional icons that provide
-          optional functionalities such as add new file, add new
-          folder, etc.
-        */}
-        <div className={[styles.nodeOptions, theme.sidebarIcon].join(' ')}>
-          {
-            nodeFolder && (
-              <div>
-                <FontAwesomeIcon
-                  className={[styles.nodeOptionsIcons, theme.sidebarOptionIcon].join(' ')}
-                  icon='folder-plus'
-                  size='xs'
-                  onClick={(e) => {
-                    this.addNewFolder(e, node.path)
-                  }}
-                />
-                <FontAwesomeIcon
-                  className={[styles.nodeOptionsIcons, theme.sidebarOptionIcon].join(' ')}
-                  icon='file-medical'
-                  size='xs'
-                  onClick={(e) => {
-                    this.addNewFile(e, node.path)
-                  }}
-                />
-              </div>
-            )
-          }
-          <div>
-            <FontAwesomeIcon
-              className={[styles.nodeOptionsIcons, theme.sidebarOptionIcon].join(' ')}
-              icon='file-signature'
-              size='xs'
-              onClick={(e) => {
-                this.renameFile(e, node)
-              }}
-            />
-          </div>
-        </div>
-
-      </span>
-    )
-  }
-
-  fileClick(node) {
-    this.props.selectFile(node)
-  }
-
   /*
-    Prompts are disallowed by eslint but for now
-    prompts are the simplest and easiest solution
-    for gathering user input.
-    That is why window.prompt lines of code are
-    ignored by eslint.
+    This method is used to render individual tree items.
+    Tree item contents are defined with SidebarTreeItem
+    component.
   */
-  addNewFolder(e, path) {
-    e.stopPropagation()
-    // eslint-disable-next-line
-    let name = window.prompt('Please enter new folder name', 'new-folder')
-    if (name === null || name === '') {
-      return
-    }
-    let newFolder = {
-      module: name
-    }
-    this.props.createFile(newFolder, path)
-  }
-
-  addNewFile(e, path) {
-    e.stopPropagation()
-    // eslint-disable-next-line
-    let name = window.prompt('Please enter new file name', 'new-file')
-    if (name === null || name === '') {
-      return
-    }
-    let newFile = {
-      module: name,
-      leaf: true
-    }
-    this.props.createFile(newFile, path)
-  }
-
-  renameFile(e, file) {
-    e.stopPropagation()
-    // eslint-disable-next-line
-    let newName = window.prompt('Please enter new file name', file.module)
-    if (newName === null || newName === '') {
-      return
-    }
-    this.props.renameFile(file, newName)
+  renderNode = node => {
+    return (<SidebarTreeItem node={node} />)
   }
 
   render() {
@@ -173,8 +50,6 @@ class SidebarTree extends Component {
 SidebarTree.propTypes = {
   files: PropTypes.object,
   selectFile: PropTypes.func,
-  createFile: PropTypes.func,
-  renameFile: PropTypes.func,
   setExamples: PropTypes.func
 }
 
@@ -187,8 +62,6 @@ function mapStateToProps(state) {
 function mapDispatchToProps(dispatch) {
   return bindActionCreators({
     selectFile: selectFile,
-    createFile: addFile,
-    renameFile: renameFile,
     setExamples: setExamples
   }, dispatch)
 }
